Move value functions to new file
resolves a circular dependency between session & lib
