+++ src/post/autumnaton.ts
@@ -1,4 +1,4 @@
-import { garboAverageValue, garboValue } from "../session";
+import { garboAverageValue, garboValue } from "../value";
 import { estimatedGarboTurns, estimatedTurnsTomorrow } from "../turns";
 import { appearanceRates, availableAmount, getMonsters, itemDropsArray, Location } from "kolmafia";
 import { $items, AutumnAton, get, maxBy, sum } from "libram";
