+++ src/tasks/dailyItems.ts
@@ -42,10 +42,10 @@ import {
 import { acquire } from "../acquire";
 import { globalOptions } from "../config";
 import { embezzlerCount } from "../embezzler";
-import { rufusPotion } from "../potions";
 import { doingExtrovermectin } from "../extrovermectin";
 import { coinmasterPrice } from "../lib";
-import { garboAverageValue, garboValue } from "../session";
+import { rufusPotion } from "../potions";
+import { garboAverageValue, garboValue } from "../value";
 
 const SummonTomes = $skills`Summon Snowcones, Summon Stickers, Summon Sugar Sheets, Summon Rad Libs, Summon Smithsness`;
 const Wads = $items`twinkly wad, cold wad, stench wad, hot wad, sleaze wad, spooky wad`;
