+++ src/yachtzee/diet.ts
@@ -40,7 +40,7 @@ import { acquire } from "../acquire";
 import { globalOptions } from "../config";
 import { hasMonsterReplacers } from "../extrovermectin";
 import { Potion } from "../potions";
-import { garboValue } from "../session";
+import { garboValue } from "../value";
 import synthesize from "../synthesis";
 import { estimatedGarboTurns } from "../turns";
 import { yachtzeePotionProfits, yachtzeePotionSetup } from "./buffs";
